+++ Game/src/com/roget/GameObject.java
@@ -4,11 +4,11 @@ import java.awt.Graphics;
 import java.awt.Rectangle;
 
 public abstract class GameObject {
-	protected int x, y;
+	protected float x, y;
 	protected ID id;
-	protected int velX, velY;
+	protected float velX, velY;
 	
-	public GameObject(int x, int y, ID id)
+	public GameObject(float x, float y, ID id)
 	{
 		this.x = x;
 		this.y = y;
@@ -29,12 +29,12 @@ public abstract class GameObject {
 		this.y = y;
 	}
 	
-	public int getX()
+	public float getX()
 	{
 		return x;
 	}
 	
-	public int getY()
+	public float getY()
 	{
 		return y;
 	}
@@ -59,12 +59,12 @@ public abstract class GameObject {
 		this.velY = velY;
 	}
 	
-	public int getVelX()
+	public float getVelX()
 	{
 		return velX;
 	}
 	
-	public int getVelY()
+	public float getVelY()
 	{
 		return velY;
 	}
